+++ services/core-go/.zig-cache/tmp/5173d82d462db4a1-cgo-gcc-input-2883778219.obj.d
@@ -1,0 +1,10 @@
+cgo-gcc-input-2883778219.o: /tmp/cgo-gcc-input-2883778219.c \
+  /opt/zig/lib/include/stddef.h \
+  /opt/zig/lib/libc/include/any-windows-any/stddef.h \
+  /opt/zig/lib/include/__stddef_header_macro.h \
+  /opt/zig/lib/include/__stddef_ptrdiff_t.h \
+  /opt/zig/lib/include/__stddef_size_t.h \
+  /opt/zig/lib/include/__stddef_wchar_t.h \
+  /opt/zig/lib/include/__stddef_null.h \
+  /opt/zig/lib/include/__stddef_max_align_t.h \
+  /opt/zig/lib/include/__stddef_offsetof.h