+++ services/core-go/.zig-cache/tmp/74972e4fa24043de-cgo-gcc-input-3803775592.o.d
@@ -1,0 +1,9 @@
+.zig-cache\tmp\74972e4fa24043de-cgo-gcc-input-3803775592.o: \
+  C:/Users/muril/AppData/Local/Temp/cgo-gcc-input-3803775592.c \
+  C:/Zig/lib/include/stddef.h C:/Zig/lib/include/__stddef_header_macro.h \
+  C:/Zig/lib/include/__stddef_ptrdiff_t.h \
+  C:/Zig/lib/include/__stddef_size_t.h \
+  C:/Zig/lib/include/__stddef_wchar_t.h \
+  C:/Zig/lib/include/__stddef_null.h \
+  C:/Zig/lib/include/__stddef_max_align_t.h \
+  C:/Zig/lib/include/__stddef_offsetof.h