+++ services/core-go/.zig-cache/tmp/95c227b060e84d21-cgo-gcc-input-1522880314.obj.d
@@ -1,0 +1,25 @@
+.zig-cache/tmp/95c227b060e84d21-cgo-gcc-input-1522880314.obj: \
+  /tmp/cgo-gcc-input-1522880314.c /opt/zig/lib/include/stddef.h \
+  /opt/zig/lib/libc/include/any-windows-any/stddef.h \
+  /opt/zig/lib/include/__stddef_header_macro.h \
+  /opt/zig/lib/include/__stddef_ptrdiff_t.h \
+  /opt/zig/lib/include/__stddef_size_t.h \
+  /opt/zig/lib/include/__stddef_wchar_t.h \
+  /opt/zig/lib/include/__stddef_null.h \
+  /opt/zig/lib/include/__stddef_max_align_t.h \
+  /opt/zig/lib/include/__stddef_offsetof.h \
+  /opt/zig/lib/libc/include/any-windows-any/stdlib.h \
+  /opt/zig/lib/libc/include/any-windows-any/corecrt.h \
+  /opt/zig/lib/libc/include/any-windows-any/_mingw.h \
+  /opt/zig/lib/libc/include/any-windows-any/_mingw_mac.h \
+  /opt/zig/lib/libc/include/any-windows-any/_mingw_secapi.h \
+  /opt/zig/lib/include/vadefs.h \
+  /opt/zig/lib/libc/include/any-windows-any/vadefs.h \
+  /opt/zig/lib/libc/include/any-windows-any/sdks/_mingw_ddk.h \
+  /opt/zig/lib/libc/include/any-windows-any/corecrt_wstdlib.h \
+  /opt/zig/lib/include/limits.h \
+  /opt/zig/lib/libc/include/any-windows-any/limits.h \
+  /opt/zig/lib/libc/include/any-windows-any/crtdefs.h \
+  /opt/zig/lib/libc/include/any-windows-any/sec_api/stdlib_s.h \
+  /opt/zig/lib/libc/include/any-windows-any/malloc.h \
+  /opt/zig/lib/include/mm_malloc.h