+++ services/core-go/.zig-cache/tmp/ab832088b0fbe799-cgo-gcc-input-3125562803.obj.d
@@ -1,0 +1,11 @@
+.zig-cache\tmp\ab832088b0fbe799-cgo-gcc-input-3125562803.obj: \
+  C:/Users/muril/AppData/Local/Temp/cgo-gcc-input-3125562803.c \
+  C:/Zig/lib/include/stddef.h \
+  C:/Zig/lib/libc/include/any-windows-any/stddef.h \
+  C:/Zig/lib/include/__stddef_header_macro.h \
+  C:/Zig/lib/include/__stddef_ptrdiff_t.h \
+  C:/Zig/lib/include/__stddef_size_t.h \
+  C:/Zig/lib/include/__stddef_wchar_t.h \
+  C:/Zig/lib/include/__stddef_null.h \
+  C:/Zig/lib/include/__stddef_max_align_t.h \
+  C:/Zig/lib/include/__stddef_offsetof.h